+++ src/main/kotlin/uz/duke/plugin/assets/DukeClips.kt
@@ -1,4 +1,4 @@
-package uz.duke.plugin.ini
+package uz.duke.plugin.assets
 
 import com.intellij.openapi.vfs.VirtualFile
 import java.io.InputStream